// client/src/containers/B2BSalaryAccount/constant.ts
/**
* @file Cognizant Salary Account Page Constant
*/
// eslint-disable-next-line no-shadow
enum SectionType {
    Button = 'button',
    UnequalCardSectionWithSubText = 'unequal_card_section_with_sub_text',
    FeatureCard = 'feature_card',
    FeatureTable = 'feature_table',
    ComingSoonCardSection = 'coming_soon_card_section',
    SingleCardSection = 'single_card_section',
    AccountOpeningSection = 'account_opening_section',
    TeamSection = 'team_section',
    FiCardSection = 'fi_card_section'
}

export { SectionType };